# apps/backend/openapi.yaml
openapi: 3.1.0
info:
  title: OpenCloudTouch
  description: Open-Source replacement for discontinued streaming device cloud features
  version: 0.2.0
paths:
...
  /api/devices/{device_id}:
    get:
      tags:
      - Devices
      summary: Get Device
      description: "Get single device by device_id.\n\nArgs:\n    device_id: Device\
        \ ID\n\nReturns:\n    Device details\n\nRaises:\n    DeviceNotFoundError:\
        \ If device does not exist"
      operationId: get_device_api_devices__device_id__get
      parameters:
      - name: device_id
        in: path
        required: true
        schema:
          type: string
          title: Device Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/devices/{device_id}/capabilities:
    get:
      tags:
      - Devices
      summary: Get Device Capabilities Endpoint
      description: "Get device capabilities for UI feature detection.\n\nReturns which\
        \ features this specific device supports:\n- HDMI control (ST300 only)\n-\
        \ Bass/balance controls\n- Available input sources\n- Zone/group support\n\
        - All supported endpoints\n\nArgs:\n    device_id: Device ID\n\nReturns:\n\
        \    Feature flags and capabilities for UI rendering\n\nExample Response:\n\
        \    {\n        \"device_id\": \"AABBCC112233\",\n        \"device_type\"\
        : \"SoundTouch 30 Series III\",\n        \"is_soundbar\": false,\n       \
        \ \"features\": {\n            \"hdmi_control\": false,\n            \"bass_control\"\
        : true,\n            \"bluetooth\": true,\n            ...\n        },\n \
        \       \"sources\": [\"BLUETOOTH\", \"AUX\", \"INTERNET_RADIO\"],\n     \
        \   \"advanced\": {...}\n    }"
      operationId: get_device_capabilities_endpoint_api_devices__device_id__capabilities_get
      parameters:
      - name: device_id
        in: path
        required: true
        schema:
          type: string
          title: Device Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/devices/{device_id}/key:
    post:
      tags:
      - Devices
      summary: Press Key
      description: "Simulate a key press on a device.\n\nUsed for E2E testing to trigger\
        \ preset playback without physical button press.\n\nArgs:\n    device_id:\
        \ Device ID\n    key: Key name (e.g., \"PRESET_1\", \"PRESET_2\", \"PRESET_3\"\
        , ...)\n    state: Key state (\"press\", \"release\", or \"both\"). Default:\
        \ \"both\"\n\nReturns:\n    Success message\n\nRaises:\n    DeviceNotFoundError:\
        \ If device does not exist\n    HTTPException(400): If key or state is invalid\n\
        \    HTTPException(500): If key press fails\n\nExample:\n    POST /api/devices/AABBCC112233/key?key=PRESET_1&state=both"
      operationId: press_key_api_devices__device_id__key_post
      parameters:
      - name: device_id
        in: path
        required: true
        schema:
          type: string
          title: Device Id
      - name: key
        in: query
        required: true
        schema:
          type: string
          title: Key
      - name: state
        in: query
        required: false
        schema:
          type: string
          default: both
          title: State
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/devices/{device_id}/now-playing:
    get:
      tags:
      - Devices
      summary: Get Now Playing
      description: Get current playback status for a device.
      operationId: get_now_playing_api_devices__device_id__now_playing_get
      parameters:
      - name: device_id
        in: path
        required: true
        schema:
          type: string
          title: Device Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/devices/{device_id}/volume:
    get:
      tags:
      - Devices
      summary: Get Volume
      description: Get current volume state for a device.
      operationId: get_volume_api_devices__device_id__volume_get
      parameters:
      - name: device_id
        in: path
        required: true
        schema:
          type: string
          title: Device Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Devices
      summary: Set Volume
      description: Set volume level (0-100).
      operationId: set_volume_api_devices__device_id__volume_put
      parameters:
      - name: device_id
        in: path
        required: true
        schema:
          type: string
          title: Device Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Body_set_volume_api_devices__device_id__volume_put'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/devices/{device_id}/mute:
    put:
      tags:
      - Devices
      summary: Set Mute
      description: Set mute state.
      operationId: set_mute_api_devices__device_id__mute_put
      parameters:
      - name: device_id
        in: path
        required: true
        schema:
          type: string
          title: Device Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Body_set_mute_api_devices__device_id__mute_put'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_set_mute_api_devices__device_id__mute_put:
      properties:
        muted:
          type: boolean
          title: Muted
      type: object
      required:
      - muted
      title: Body_set_mute_api_devices__device_id__mute_put
    Body_set_volume_api_devices__device_id__volume_put:
      properties:
        level:
          type: integer
          maximum: 100.0
          minimum: 0.0
          title: Level
      type: object
      required:
      - level
      title: Body_set_volume_api_devices__device_id__volume_put
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...